end-of-week commit (not working)
