+++ public/src/components/amounts/DeleteTestDialog.tsx
