+++ public/src/components/amounts/DeleteVariantButton.tsx
