+++ public/src/components/amounts/DeleteVariantDialog.tsx
