+++ public/src/components/amounts/old_configuredAmountsEditor.tsx
@@ -1,98 +1,3 @@
-import React, { useState } from 'react';
-import { 
-  Territories,
-  Territory,
-  AmountsTests,
-  AmountsTest,
-  Regions,
-  Countries,
-} from '../../utils/models';
-
-import AmountsTestsList from './AmountsTestsList';
-import AmountsTestEditor from './AmountsTestEditor';
-
-import {
-  SupportFrontendSettingsType,
-  fetchSupportFrontendSettings,
-  saveSupportFrontendSettings,
-} from '../../utils/requests';
-import withS3Data, { InnerProps, DataFromServer } from '../../hocs/withS3Data';
-
-const AmountsForm: React.FC<InnerProps<AmountsTests>> = ({
-  data,
-  setData,
-  saveData,
-  saving: boolean,
-}: InnerProps<AmountsTests>) => {
-
-  // const [currentTests, setCurrentTests] = useState(data || []);
-  const [selectedTest, setSelectedTest] = useState<AmountsTest | undefined>();
-  // const [saving, setSaving] = useState(false);
-  // const [creating, setCreating] = useState(false);
-
-  const onTargetSelected = (target: Territory) => {
-    const currentTest = data.filter(test => test.target === target);
-    if (currentTest.length) {
-      setSelectedTest(currentTest[0]);
-    }
-    else {
-      setSelectedTest(undefined);
-    }
-  };
-
-  const saveTest = (updated: AmountsTest) => {
-    const updatedTests = data.filter(t => t.testName !== updated.testName);
-    updatedTests.push(updated);
-    saveData(updatedTests);
-  };
-
-  const createTest = () => {
-
-  };
-
-  const deleteTest = (test: AmountsTest) => {
-    const updatedTests = data.filter(t => t.testName !== test.testName);
-    saveData(updatedTests);
-  };
-
-  return (
-    <div>
-      <AmountsTestsList 
-        tests={data}
-        selectedTest={selectedTest?.testName || undefined}
-        onTargetSelected={onTargetSelected}
-        save={saveTest}
-        saving={saving}
-        create={createTest}
-        creating={creating}
-      />
-      <AmountsTestEditor 
-        test={selectedTest}
-        updateTest={saveTest}
-        deleteTest={deleteTest}
-      />
-    </div>
-  );
-};
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const fetchSettings = (): Promise<any> =>
-  fetchSupportFrontendSettings(SupportFrontendSettingsType.amounts);
-
-const saveSettings = (data: DataFromServer<AmountsTests>): Promise<Response> =>
-  saveSupportFrontendSettings(SupportFrontendSettingsType.amounts, data);
-
-export default withS3Data<AmountsTests>(AmountsForm, fetchSettings, saveSettings);
-
-
-
-
-
-
-
-
-
-
 // import React, { useState } from 'react';
 
 // import { makeStyles, Theme } from '@material-ui/core/styles';
